purger: remove old exports

+++ wazo_call_logd/purger.py
@@ -10,6 +10,7 @@ from sqlalchemy import func
 from .database.models import (
     CallLog,
     Config,
+    Export,
     Recording,
     Retention,
     Tenant,
@@ -19,6 +20,15 @@ logger = logging.getLogger(__name__)
 
 DEFAULT_PURGE_DB_DAYS = 365
 DEFAULT_CALL_LOGD_DAYS = 365
+DEFAULT_CALL_LOGD_EXPORT_DAYS = 2
+
+
+def _remove_export_files(exports):
+    for export in exports:
+        try:
+            os.remove(export.path)
+        except FileNotFoundError:
+            logger.info('Export file already deleted: "%s"')
 
 
 def _remove_recording_files(call_logs):
@@ -36,16 +46,20 @@ def _remove_recording_files(call_logs):
                     )
 
 
-def _extract_days_to_keep(tenant_days, default_days, purge_db_days):
+def _extract_days_to_keep(tenant_days, default_days, purge_db_days, default=DEFAULT_CALL_LOGD_DAYS):
     if tenant_days is not None:
         return tenant_days
 
     purge_db_days_customized = purge_db_days != DEFAULT_PURGE_DB_DAYS
-    default_days_customized = default_days != DEFAULT_CALL_LOGD_DAYS
+    default_days_customized = default_days != default
     if purge_db_days_customized and not default_days_customized:
         return purge_db_days
 
     return default_days
+
+
+def _extract_days_to_keep_for_exports(*args):
+    return _extract_days_to_keep(*args, default=DEFAULT_CALL_LOGD_EXPORT_DAYS)
 
 
 class CallLogsPurger:
@@ -77,6 +91,33 @@ class CallLogsPurger:
                 .filter(CallLog.date < max_date)
                 .filter(CallLog.tenant_uuid == tenant.uuid)
             )
+            query.delete(synchronize_session=False)
+
+
+class ExportsPurger:
+    def purge(self, days_to_keep, session):
+        config = session.query(Config).first()
+        if not config:
+            raise Exception('No default config found')
+        default_days = config.retention_export_days
+
+        retentions = {r.tenant_uuid: r for r in session.query(Retention).all()}
+        tenants = session.query(Tenant).all()
+        for tenant in tenants:
+            days = days_to_keep
+            retention = retentions.get(tenant.uuid)
+            tenant_days = retention.export_days if retention else None
+            days = _extract_days_to_keep_for_exports(tenant_days, default_days, days_to_keep)
+
+            max_date = func.now() - datetime.timedelta(days=days)
+            query = (
+                session.query(Export)
+                .filter(Export.requested_at < max_date)
+                .filter(Export.tenant_uuid == tenant.uuid)
+            )
+            exports = query.all()
+            _remove_export_files(exports)
+
             query.delete(synchronize_session=False)
